Promote PromQL evaluator values to double precision
Keep gauge scans float-backed at the storage boundary while carrying scalar literals and evaluated vector samples as doubles. Preserve regression math precision through deriv and predict_linear, update REST scalar serialization, and adapt test helpers to the evaluator-domain type.

+++ okapi-promql/src/main/java/org/okapi/promql/eval/ScalarResult.java
@@ -8,9 +8,9 @@ import lombok.Getter;
 
 // eval/ScalarResult.java
 public final class ScalarResult implements ExpressionResult {
-  @Getter public final float value;
+  @Getter public final double value;
 
-  public ScalarResult(float value) {
+  public ScalarResult(double value) {
     this.value = value;
   }
