+++ okapi-promql/src/main/java/org/okapi/promql/eval/VectorData.java
@@ -16,12 +16,12 @@ public class VectorData {
     }
   }
 
-  public record Sample(long ts, long sourceTs, float value, HistogramSeries.HistogramSample histogram) {
-    public Sample(long ts, float value) {
+  public record Sample(long ts, long sourceTs, double value, HistogramSeries.HistogramSample histogram) {
+    public Sample(long ts, double value) {
       this(ts, ts, value, null);
     }
 
-    public Sample(long ts, long sourceTs, float value) {
+    public Sample(long ts, long sourceTs, double value) {
       this(ts, sourceTs, value, null);
     }
